Refactored code structure to match new UML

+++ src/main/java/com/example/sysc4806project/Book.java
@@ -54,7 +54,7 @@ public class Book {
         this.price = price;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
